Start with Analyzer in Java

+++ decompile/decompile.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 clear
 
+if [ "$#" -ne 2 ]; then
+	echo "Usage: ./decompile.sh input_path output_path"
+	exit
+fi
+
 INPUT=$1
 OUTPUT=$2
 
@@ -8,23 +13,31 @@ OUTPUT=$2
 for FILE in $INPUT/*.apk; do
 
 	FOLDER=${FILE/.apk/""}
-	FOLDER=$(echo $FOLDER| cut -c 3-100000)
+	START=${#INPUT}
+	START=$((START+2))
+	FOLDER=$(echo $FOLDER| cut -c ${START}-100000)
+
+	# check if folder exists
+	if [ -d "$OUTPUT/$FOLDER" ]; then
+		continue
+	fi
+
 	echo $FOLDER
 
 	# create output folder for decompiled classes
-	rm -R "$OUTPUT/$FOLDER"
+	#rm -R "$OUTPUT/$FOLDER"
 	mkdir -p "$OUTPUT/$FOLDER"
 
 	# extract .jar from .apk
-	dex2jar/d2j-dex2jar.sh -o "$OUTPUT/$FOLDER.jar" "$FOLDER.apk"
+	dex2jar/d2j-dex2jar.sh -o "$OUTPUT/$FOLDER.jar" "$INPUT/$FOLDER.apk"
 
 	# extract resources from apk files
-	java -jar apktool/apktool_2.0.3.jar decode -f "$FOLDER.apk" -o "$OUTPUT/$FOLDER/"
+	java -jar apktool/apktool_2.0.3.jar decode -f "$INPUT/$FOLDER.apk" -o "$OUTPUT/$FOLDER/"
 	rm -Rf "$OUTPUT/$FOLDER/smali"
 	rm "$OUTPUT/$FOLDER/apktool.yml"
 
 	# decompile jar archive
-	java -jar fernflower/fernflower.jar -ren=1 "$OUTPUT/$FOLDER.jar" "$OUTPUT/$FOLDER/"
+	java -jar fernflower/fernflower.jar -log=ERROR -dgs=1 -ren=1 "$OUTPUT/$FOLDER.jar" "$OUTPUT/$FOLDER/"
 	unzip "$OUTPUT/$FOLDER/$FOLDER.jar" -d "$OUTPUT/$FOLDER/"
 	rm "$OUTPUT/$FOLDER.jar" 
 	rm "$OUTPUT/$FOLDER/$FOLDER.jar"
